# packages/shared/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "WebAuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, tokens set in httpOnly cookies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Login with email and password",
        "tags": [
          "Web Authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "WebAuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token refreshed successfully, new tokens set in cookies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "summary": "Refresh access token using refresh token",
        "tags": [
          "Web Authentication"
        ]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "operationId": "WebAuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "summary": "Get current authenticated user",
        "tags": [
          "Web Authentication"
        ]
      }
    },
    "/api/v1/mcp/servers": {
      "post": {
        "operationId": "McpRegistryController_createServer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Server created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Server with providedId already exists"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Register a new MCP server",
        "tags": [
          "MCP Registry"
        ]
      },
      "get": {
        "operationId": "McpRegistryController_listServers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of servers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "List all MCP servers with pagination",
        "tags": [
          "MCP Registry"
        ]
      }
    },
    "/api/v1/mcp/servers/{serverId}": {
      "get": {
        "operationId": "McpRegistryController_getServer",
        "parameters": [
          {
            "name": "serverId",
            "required": true,
            "in": "path",
            "description": "Server UUID or provided ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Server not found"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Get MCP server by UUID or provided ID",
        "tags": [
          "MCP Registry"
        ]
      },
      "patch": {
        "operationId": "McpRegistryController_updateServer",
        "parameters": [
          {
            "name": "serverId",
            "required": true,
            "in": "path",
            "description": "Server UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Server not found"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Update MCP server details",
        "tags": [
          "MCP Registry"
        ]
      },
      "delete": {
        "operationId": "McpRegistryController_deleteServer",
        "parameters": [
          {
            "name": "serverId",
            "required": true,
            "in": "path",
            "description": "Server UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteServerResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Server not found"
          },
          "409": {
            "description": "Server has dependencies"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Delete MCP server (must have no dependencies)",
        "tags": [
          "MCP Registry"
        ]
      }
    },
...
    "/.well-known/oauth-authorization-server/mcp/{mcpServerId}/{version}": {
      "get": {
        "description": "Provides discovery metadata (RFC 8414) for OAuth 2.0 clients integrating with a specific MCP server version. Accepts either the server UUID or the providedId.",
        "operationId": "DiscoveryController_getAuthorizationServerMetadata",
        "parameters": [
          {
            "name": "mcpServerId",
            "required": true,
            "in": "path",
            "description": "MCP server UUID or providedId",
            "schema": {
              "example": "inventory-service",
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "description": "Semantic version of the MCP server",
            "schema": {
              "example": "v1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorization server metadata retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationServerMetadataDto"
                }
              }
            }
          }
        },
        "summary": "Expose OAuth 2.0 Authorization Server metadata for a registered MCP server version",
        "tags": [
          "Discovery"
        ]
      }
    },
    "/.well-known/oauth-protected-resource/{path}": {
      "get": {
        "operationId": "DiscoveryController_all",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "path",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Discovery"
        ]
      }
    },
...
  },
  "info": {
    "title": "AI Monorepo API",
    "description": "The AI Monorepo API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-Cookie": {
        "type": "http",
        "in": "cookie",
        "name": "access_token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          },
          "displayName": {
            "type": "string",
            "description": "User display name",
            "example": "John Doe"
          },
          "role": {
            "type": "string",
            "description": "User role",
            "enum": [
              "admin",
              "standard"
            ],
            "example": "standard"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "role"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "description": "Authenticated user information",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserResponseDto"
              }
            ]
          },
          "expiresIn": {
            "type": "number",
            "description": "Access token expiration time in seconds",
            "example": 600
          }
        },
        "required": [
          "user",
          "expiresIn"
        ]
      },
      "CreateServerDto": {
        "type": "object",
        "properties": {
          "providedId": {
            "type": "string",
            "description": "Human-readable unique identifier for the MCP server",
            "example": "github-integration"
          },
          "name": {
            "type": "string",
            "description": "Display name of the MCP server",
            "example": "GitHub Integration"
          },
          "description": {
            "type": "string",
            "description": "Short description of the MCP server",
            "example": "Provides access to GitHub repositories and issues"
          },
          "url": {
            "type": "string",
            "description": "URL that MCP Clients will use to connect to the server",
            "example": "http://localhost:3000/api/v1/taskeroo/tasks/mcp"
          }
        },
        "required": [
          "providedId",
          "name",
          "description"
        ]
      },
      "ServerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "System-generated UUID for the MCP server",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "providedId": {
            "type": "string",
            "description": "Human-readable unique identifier for the MCP server",
            "example": "github-integration"
          },
          "name": {
            "type": "string",
            "description": "Display name of the MCP server",
            "example": "GitHub Integration"
          },
          "description": {
            "type": "string",
            "description": "Short description of the MCP server",
            "example": "Provides access to GitHub repositories and issues"
          },
          "url": {
            "type": "string",
            "description": "URL that MCP Clients will use to connect to the server",
            "example": "http://localhost:3000/api/v1/taskeroo/tasks/mcp"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when the server was created",
            "example": "2025-11-05T08:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when the server was last updated",
            "example": "2025-11-05T08:00:00.000Z"
          }
        },
        "required": [
          "id",
          "providedId",
          "name",
          "description",
          "createdAt",
          "updatedAt"
        ]
      },
      "ServerListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "List of MCP servers",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerResponseDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of servers",
            "example": 100
          },
          "page": {
            "type": "number",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Number of items per page",
            "example": 50
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ]
      },
      "UpdateServerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the MCP server",
            "example": "GitHub Integration"
          },
          "description": {
            "type": "string",
            "description": "Short description of the MCP server",
            "example": "Provides access to GitHub repositories and issues"
          },
          "url": {
            "type": "string",
            "description": "URL that MCP Clients will use to connect to the server",
            "example": "http://localhost:3000/api/v1/taskeroo/tasks/mcp"
          }
        }
      },
      "DeleteServerResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Confirmation message indicating the server was deleted",
            "example": "Server deleted successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "AuthorizationServerMetadataDto": {
        "type": "object",
        "properties": {
          "issuer": {
            "type": "string",
            "description": "Issuer identifier for the MCP authorization server",
            "example": "https://auth.example.com"
          },
          "authorization_endpoint": {
            "type": "string",
            "description": "Authorization endpoint for initiating OAuth 2.0 authorization code flows",
            "example": "https://auth.example.com/api/v1/authorize/mcp/inventory/v1"
          },
          "token_endpoint": {
            "type": "string",
            "description": "Token endpoint for exchanging OAuth 2.0 authorization codes",
            "example": "https://auth.example.com/api/v1/token/mcp/inventory/v1"
          },
          "registration_endpoint": {
            "type": "string",
            "description": "Dynamic client registration endpoint for MCP integrations",
            "example": "https://auth.example.com/api/v1/register/mcp/inventory/v1"
          },
          "scopes_supported": {
            "description": "Scopes supported by this MCP authorization server",
            "example": [
              "mcp.read",
              "mcp.write",
              "openid",
              "offline_access"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "response_types_supported": {
            "description": "OAuth 2.0 response types supported by this authorization server",
            "example": [
              "code"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "grant_types_supported": {
            "type": "array",
            "description": "OAuth 2.0 grant types supported by this authorization server",
            "example": [
              "authorization_code",
              "refresh_token"
            ],
            "items": {
              "type": "string",
              "enum": [
                "authorization_code",
                "refresh_token"
              ]
            }
          },
          "token_endpoint_auth_methods_supported": {
            "description": "Client authentication methods supported by the token endpoint",
            "example": [
              "client_secret_basic",
              "private_key_jwt"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "code_challenge_methods_supported": {
            "description": "PKCE code challenge methods supported by this authorization server",
            "example": [
              "S256"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "issuer",
          "authorization_endpoint",
          "token_endpoint",
          "registration_endpoint",
          "scopes_supported",
          "response_types_supported",
          "grant_types_supported",
          "token_endpoint_auth_methods_supported",
          "code_challenge_methods_supported"
        ]
      },
...
    }
  }
}